make 6-size work for 32bit and 64bit long long byte

+++ 0x00-hello_world/6-size.c
@@ -9,7 +9,7 @@ int main(void)
 {
 	int i;
 	long int i2;
-	long int i3;
+	long long int i3;
 	char c;
 	float f;
